ChatBot/requirements.txt
Flask
PyPDF2
sentence-transformers
faiss-cpu
requests
